requirements.txt
Flask==2.2.5
gTTS==2.2.3
assemblyai==0.37.0
deep-translator==1.7.0
flask-cors==3.0.10
python-dotenv==0.21.0
